Incidents: Acceptiong values from ServiceNow lables instead hardcoded values
User is allowed to change more or less any drop down value that is shown
in web app. Some of those are (state, urgency, hold_reason, ..). Because
those are dynamic lists, now we pull those values from sys_choices table.
Then we map labels to values.

Thing that are not covered yet:
 - this is breaking change
 - Property status have 2 labels with additional logic. When values were
   hardcodet that was not an issue, but now changing lables inside web app
   can brake this collection:
   - "Absent":
     - in every case delete incident
     - does not care what is set for that label/value inside web app
   - "On hold"
     - user can rename this field, and we would not check for hold_reason