Splitted mappings to seperate fields.
While implementing it inside change_request, it become obvious that
we must ask for foverrides only for fields that we expect to use in
that request. Previously mapping was one object for all (incident,
change_request, ...) which made documentation totally unusable. Other
side effect was that many of subobject were not used at all.